beaglebone: drop rootfs size restriction (WIP)
This image is to be installed on large-partitioned devices only.

WIP since this needs to be handled differently. My idea is to install
it by default for machines that have been large partitioned from the
start, such as the venus-touch, rpi & easysolar & other nano-pi based
products.

And then there is the beaglebone machine, which is installed on two
different products:

1) the Venus GX. which has for a long time be partitioned with a small
   space for the rootfs. Somewhere in 2018 this was changed in
   production to a larger partition. Automatically changing the
   partitioning for devices has been decided to be a route we're not
   going to take. It is possible to reformat ofcourse; by running a
   production image and first taking out (& afterwards restoring) a
   few files such as wifi keys.

2) Octo GX. Partitioned with large partitions from day one.

https://github.com/victronenergy/venus/wiki/machines