packagegroup-venus-base: add node-red, nodejs-npm & node-red-contrib-victron (WIP)
Why wip? Because something more special must be done: Add it only to images for machines
with enough disk size, or something like that.

For now its fine, we just make swu files; and people can only install them on a
Venus GX with big enough partitions, as well as ofcourse rpis and Cerbo GX.